update logic in databases

+++ databases/user_database.py
@@ -1,18 +1,17 @@
 # user_database.py
 import sqlite3
+import os
 from werkzeug.security import generate_password_hash, check_password_hash
-
+from routes.utils import DATABASEFOLDER
+from database import BookDatabase
 class UserDatabase:
 
-    ADMIN_USERNAME = "admin"
-    ADMIN_PASSWORD = "change_this_password"
     ROLE_USER = "user"
     ROLE_ADMIN = "admin"
 
-    def __init__(self, path="users.db"):
+    def __init__(self, path="/databases/users.db"):
         self.path = path
         self._create_table()
-        self._create_default_admin()
 
     def _get_connection(self):
         conn = sqlite3.connect(self.path)
@@ -34,32 +33,20 @@ class UserDatabase:
             )
         """)        
         conn.commit()
+        conn.close()
 
-    def _create_default_admin(self):
+    def promote_user(self, user_id):
         conn = self._get_connection()
         cursor = conn.cursor()
         cursor.execute(
-            "SELECT id FROM users WHERE username=?",
-            (self.ADMIN_USERNAME,)
-        )
-        if cursor.fetchone() is None:
-            password_hash = generate_password_hash(self.ADMIN_PASSWORD)
-            cursor.execute(
-                """
-                INSERT INTO users
-                (username, password_hash, role)
-                VALUES (?, ?, ?)
-                """,
-                (
-                    self.ADMIN_USERNAME,
-                    password_hash,
-                    self.ROLE_ADMIN
-                )
+                "UPDATE users SET role=? WHERE id=?",
+                (self.ROLE_ADMIN, user_id)
             )
-            conn.commit()
+        conn.commit()
+        conn.close()
 
     def _validate_username(self, username):
-        username = username.strip()
+        username = username.strip().lower()
         if not username:
             raise ValueError("Username cannot be empty.")
         if len(username) < 3:
@@ -69,7 +56,6 @@ class UserDatabase:
         return username
     
     def _validate_password(self, password):
-        password = password.strip()
         if not password:
             raise ValueError("Password cannot be empty.")
         if len(password) < 8:
@@ -88,18 +74,23 @@ class UserDatabase:
                 (username, password_hash, self.ROLE_USER)
             )
             conn.commit()
-            return True # new username
+            conn.close()
+            BookDatabase(DATABASEFOLDER, user_id)
+
+            return cursor.lastrowid
         except sqlite3.IntegrityError as e:
             conn.rollback()
+            conn.close()
             if "UNIQUE constraint failed: users.username" in str(e):
                 return False
             raise
 
-    def verify_user(self, username, password):
+    def verify_user(self, user_id, password):
         conn = self._get_connection()
         c = conn.cursor()
-        c.execute("SELECT password_hash FROM users WHERE username = ?", (username,))
+        c.execute("SELECT password_hash FROM users WHERE id = ?", (user_id,))
         row = c.fetchone()
+        conn.close()
         if row is None:
             return False
 
@@ -108,18 +99,35 @@ class UserDatabase:
             password
         )
 
-    def get_role(self, username):
+    def login_user(self, username, password):
+        username = self._validate_username(username)
         conn = self._get_connection()
         c = conn.cursor()
-        c.execute("SELECT role FROM users WHERE username = ?", (username,))
+        c.execute("SELECT id, password_hash FROM users WHERE username=?", (username,))
         row = c.fetchone()
+        conn.close()
+        if row is None:
+            return False
+        if check_password_hash(row["password_hash"], password):
+            return True
+        return False
+
+    def get_role(self, user_id):
+        conn = self._get_connection()
+        c = conn.cursor()
+        c.execute("SELECT role FROM users WHERE id = ?", (user_id,))
+        row = c.fetchone()
+        conn.close()
         return row["role"] if row else None
         
     def close_database(self):
         conn = self._get_connection()
         conn.close()
 
-    def change_password(self, username, new_password):
+    def change_password(self, user_id, password, new_password):
+        if not self.verify_user(user_id, password):
+            return False
+        new_password = self._validate_password(new_password)
         new_hash = generate_password_hash(new_password)
         conn = self._get_connection()
         cursor = conn.cursor()
@@ -127,27 +135,70 @@ class UserDatabase:
             """
             UPDATE users
             SET password_hash=?
-            WHERE username=?
+            WHERE id=?
             """,
             (
                 new_hash,
-                username
+                user_id, 
             )
         )
         conn.commit()
         conn.close()
+        return True
 
-    def delete_user(self, username):
+    def change_username(self, user_id, username, new_username):
+        new_username = self._validate_username(new_username)
         conn = self._get_connection()
         cursor = conn.cursor()
+        try:
+            cursor.execute(
+                """
+                UPDATE users
+                SET username=?
+                WHERE id=?
+                """,
+                (
+                    new_username,
+                    user_id
+                )
+            )
+            conn.commit()
+            conn.close()
+            return True
+        except sqlite3.IntegrityError as e:
+            conn.rollback()
+            conn.close()
+            if "UNIQUE constraint failed: users.username" in str(e):
+                return False
+            raise
+
+    def delete_user(self, user_id):
+        conn = self._get_connection()
+        cursor = conn.cursor()
+        cursor.execute("""DELETE FROM users WHERE id=?""",(user_id,))
+        conn.commit()
+        conn.close()
+        deleted = cursor.rowcount > 0
+        if deleted:
+            book_path = os.path.join(DATABASEFOLDER, f"books_{user_id}.db")
+            if os.path.exists(book_path):
+                os.remove(book_path)
+        return deleted
+
+    def get_user_id(self, username):
+        conn = self._get_connection()
+        cursor = conn.cursor()
+
         cursor.execute(
             """
-            DELETE FROM users
+            SELECT id
+            FROM users
             WHERE username=?
             """,
             (username,)
         )
-        conn.commit()
+
+        row = cursor.fetchone()
         conn.close()
 
-    
+        return row["id"] if row else None
